% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIbex.R
\name{filter.cells}
\alias{filter.cells}
\title{Filter Single-Cell Data Based on CDR3 Sequences}
\usage{
filter.cells(sc.obj, chain)
}
\arguments{
\item{sc.obj}{A Seurat or SingleCellExperiment object.}

\item{chain}{Character. Specifies the chain type ("Heavy" or "Light").}
}
\value{
A filtered Seurat or SingleCellExperiment object.
}
\description{
This function subsets a Seurat or SingleCellExperiment object,
removing cells where the \code{CTaa} column is missing or contains unwanted patterns.
}
