% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpandedSequences.R
\name{getExpandedSequences}
\alias{getExpandedSequences}
\title{Extract full CDR sequences from contig files}
\usage{
getExpandedSequences(contig.files, chain = "Heavy")
}
\arguments{
\item{chain}{Heavy or Light chain to extract}

\item{sc}{A list of contig files with the "cdr1", "cdr2" and "cdr3" columns}
}
\value{
A reorganized and filtered contig list with expanded_aa sequences
}
\description{
This function is to generate CDR1/2/3 expanded
sequences that are functional with the appropriate model.
}
\details{
#' @examples
x <- quietBCRgenes(ibex_example)
}
