% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietBCRgenes.R
\name{getHumanIgPseudoGenes}
\alias{getHumanIgPseudoGenes}
\title{Human Immunoglobulin pseudogenes}
\usage{
getHumanIgPseudoGenes()
}
\value{
Character vector of human immunoglobulin pseudogenes.
}
\description{
This function returns a character vector of human immunoglobulin
pseudogenes. These are also the genes that are removed from the
variable gene list in the \code{quietBCRgenes} function.
}
