% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ibex-package.R
\docType{package}
\name{Ibex-package}
\alias{Ibex}
\alias{Ibex-package}
\title{Ibex: Methods for BCR single-cell embedding}
\description{
Implementation of the Ibex algorithm for single-cell embedding based on BCR sequences. The package includes a standalone function to encode BCR sequence information by amino acid properties or sequence order using tensorflow-based autoencoder. In addition, the package interacts with SingleCellExperiment or Seurat data objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BorchLab/Ibex/}
  \item Report bugs at \url{https://github.com/BorchLab/Ibex/issues}
}

}
\author{
\strong{Maintainer}: Nick Borcherding \email{ncborch@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Qile Yang \email{qile.yang@berkeley.edu} [contributor]
}

}
\keyword{internal}
