% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibex_vdj.R
\docType{data}
\name{ibex_vdj}
\alias{ibex_vdj}
\title{200 randomly sampled cells from  the 10x 
10k Human Diseased PBMCs (ALL) Freshly Processed Data}
\description{
200 randomly sampled cells from  the 10x 
10k Human Diseased PBMCs (ALL) Freshly Processed Data
}
