% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietBCRgenes.R
\name{quietBCRgenes}
\alias{quietBCRgenes}
\title{Remove BCR Genes from Variable Gene Results}
\usage{
quietBCRgenes(sc, assay = NULL)
}
\arguments{
\item{sc}{A single-cell dataset, which can be:
\itemize{
  \item A Seurat object
  \item A vector of variable genes generated by workflows such as Bioconductor's \code{scran}
}}

\item{assay}{Character. Specifies the Seurat assay slot to use for removing BCR genes.
If \code{NULL}, the function defaults to the active assay in the Seurat object.}
}
\value{
The input Seurat object or vector with BCR genes removed from the variable features.
}
\description{
This function removes B-cell receptor (BCR) genes from the variable features of a 
single-cell dataset. Most single-cell workflows prioritize highly expressed and 
highly variable genes for principal component analysis (PCA) and dimensional 
reduction. By excluding BCR genes, this function ensures that the variable gene 
set focuses on biologically relevant features rather than highly variable BCR genes.
}
\examples{
# Remove BCR genes from the variable features of a Seurat object
ibex_example <- quietBCRgenes(ibex_example)

}
