% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIbex.R
\name{runIbex}
\alias{runIbex}
\title{Ibex single cell calculation}
\usage{
runIbex(sc, chains = "Heavy", AA.properties = "AF", reduction.name = "Ibex")
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{Heavy or Light}

\item{AA.properties}{Amino acid properties to use for distance calculation: 
"AF" = Atchley factors, "KF" = Kidera factors, "both" = AF and KF, or "OHE" for
One Hot Autoencoder}

\item{reduction.name}{Keyword to save Ibex reduction. Useful if you want
to try Ibex with multiple parameters}
}
\value{
Seurat or SingleCellExperiment object with Ibex dimensions placed 
into the dimensional reduction slot.
}
\description{
Run Ibex algorithm with Seurat or SingleCellExperiment pipelines
}
\examples{
ibex_example <- runIbex(ibex_example, 
                        AA.properties = "AF", 
                        reduction.name = "Ibex.AF")
                        
}
