% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibex.data.R
\docType{data}
\name{ibex.data}
\alias{ibex.data}
\title{Amino acid properties and ranges for
the autoencoder models along tensor}
\description{
Amino acid properties and ranges for
the autoencoder models along tensor
}
