% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIbex.R
\name{Ibex.matrix}
\alias{Ibex.matrix}
\title{Main Ibex interface}
\usage{
Ibex.matrix(
  sc,
  chains = "Heavy",
  method = "encoder",
  encoder.model = "VAE",
  encoder.input = "AF",
  theta = pi
)
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format or
the output of combineBCR() in scRepertoire}

\item{chains}{Heavy or Light}

\item{method}{"encoder" = using deep learning autoencoders or 
"geometric" = geomteric transformations based on BLOSUM62 matrix}

\item{encoder.model}{"AE" = dense autoencoder or "VAE" = variation autoencoder}

\item{encoder.input}{"AF" = Atchley factors, "KF" = Kidera factors, "both" = AF and KF, or "OHE" for
One Hot Autoencoder}

\item{theta}{angle to use for geometric transformation}
}
\value{
Ibex encoded values from the autoencoder
}
\description{
Use this to run the Ibex algorithm to return latent vectors in 
the form of a matrix or if you prefer.
}
\examples{
ibex_values <- Ibex.matrix(ibex_example, 
                          chains = "Heavy",
                          method = "encoder",
                          encoder.model = "VAE",
                          encoder.input = "AF")
                          
ibex_values <- Ibex.matrix(ibex_example, 
                          chains = "Heavy",
                          method = "geometric",
                          theta = pi)
                        
}
