% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIbex.R
\name{Ibex.matrix}
\alias{Ibex.matrix}
\title{Main Ibex interface}
\usage{
Ibex.matrix(sc, chains = "Heavy", AA.properties = "AF")
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{Heavy or Light}

\item{AA.properties}{Amino acid properties to use for distance calculation: 
"AF" = Atchley factors, "KF" = Kidera factors, "both" = AF and KF, or "OHE" for
One Hot Autoencoder}
}
\value{
Ibex encoded values from the autoencoder
}
\description{
Use this to run the Ibex algorithm to return latent vectors in 
the form of a matrix or if you prefer.
}
\examples{
ibex_values <- Ibex.matrix(ibex_example, 
                        chains = "Heavy",
                        AA.properties = "AF")
                        
}
