% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibex_example.R
\docType{data}
\name{ibex_example}
\alias{ibex_example}
\title{A seurat object of 370 single B cells in
multisystem inflammatory syndrome}
\description{
A seurat object of 370 single B cells in
multisystem inflammatory syndrome
}
