% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibex_example.R
\docType{data}
\name{ibex_example}
\alias{ibex_example}
\title{A SCE object of 200 randomly-sampled
B cells with BCR sequences from the 10x
2k_BEAM-Ab_Mouse_HEL_5pv2_2k_BEAM-Ab_Mouse_HEL_5pv2}
\description{
A SCE object of 200 randomly-sampled
B cells with BCR sequences from the 10x
2k_BEAM-Ab_Mouse_HEL_5pv2_2k_BEAM-Ab_Mouse_HEL_5pv2
}
