% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibex_vdj.R
\docType{data}
\name{ibex_vdj}
\alias{ibex_vdj}
\title{Full filtered_annotated_contig.csv from  the 10x 
2k_BEAM-Ab_Mouse_HEL_5pv2_2k_BEAM-Ab_Mouse_HEL_5pv2}
\description{
Full filtered_annotated_contig.csv from  the 10x 
2k_BEAM-Ab_Mouse_HEL_5pv2_2k_BEAM-Ab_Mouse_HEL_5pv2
}
