% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietBCRgenes.R
\name{quietBCRgenes}
\alias{quietBCRgenes}
\title{Remove BCR genes from variable gene results}
\usage{
quietBCRgenes(sc, assay = NULL)
}
\arguments{
\item{sc}{Single-cell object in Seurat format or vector of variable genes to use in reduction}

\item{assay}{The Seurat assay slot to use to remove BCR genes from, NULL value will default to
the default assay}
}
\value{
Seurat object or vector list with BCR genes removed.
}
\description{
Most single-cell workflows use highly-expressed and highly-variable
genes for the initial calculation of PCA and subsequent dimensional
reduction. This function will remove the BCR genes from the variable
features in the Seurat object or from a vector generated by
the Bioconductor scran workflow.
}
\examples{
x <- quietBCRgenes(ibex_example)

}
