% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIbex.R
\name{runIbex}
\alias{runIbex}
\title{Ibex single cell calculation}
\usage{
runIbex(
  sc,
  chains = "Heavy",
  method = "encoder",
  encoder.model = "VAE",
  encoder.input = "AF",
  theta = pi,
  reduction.name = "Ibex"
)
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{Heavy or Light}

\item{method}{"encoder" = using deep learning autoencoders or 
"geometric" = geomteric transformations based on BLOSUM62 matrix}

\item{encoder.model}{"AE" = dense autoencoder or "VAE" = variation autoencoder}

\item{encoder.input}{"AF" = Atchley factors, "KF" = Kidera factors, "both" = AF and KF, or "OHE" for
One Hot Autoencoder}

\item{theta}{angle to use for geometric transformation}

\item{reduction.name}{Keyword to save Ibex reduction. Useful if you want
to try Ibex with multiple parameters}
}
\value{
Seurat or SingleCellExperiment object with Ibex dimensions placed 
into the dimensional reduction slot.
}
\description{
Run Ibex algorithm with Seurat or SingleCellExperiment pipelines
}
\examples{
ibex_example <- runIbex(ibex_example, 
                        chains = "Heavy",
                        method = "encoder",
                        encoder.model = "VAE",
                        encoder.input = "AF")
                        
}
